# swagger_api.json
{"openapi":"3.0.1","info":{"title":"worker-manage API文档","description":"worker-manage微服务接口文档","version":"1.0"},"servers":[{"url":"http://localhost:8090","description":"Generated server url"}],"paths":{"/api/managerLogin":{"post":{"tags":["login-controller"],"summary":"管理员登录","description":"根据账号密码验证管理员身份","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/login":{"post":{"tags":["login-controller"],"summary":"员工登录","description":"根据账号密码验证员工身份","operationId":"userLogin","parameters":[{"name":"name","in":"query","required":true,"schema":{"type":"string"}},{"name":"password","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/getWorkTimeInfo":{"get":{"tags":["main-plane-controller"],"operationId":"getWorkTimeInfo","parameters":[{"name":"name","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/WorkTime"}}}}}}}},"/api/employees":{"get":{"tags":["main-plane-controller"],"operationId":"getEmployees","parameters":[{"name":"name","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Worker"}}}}}}}}},"components":{"schemas":{"WorkTime":{"type":"object","properties":{"pid":{"type":"integer","format":"int32"},"name":{"type":"string"},"goWorkTime":{"type":"string","format":"date-time"},"goHomeTime":{"type":"string","format":"date-time"},"late":{"type":"string","format":"date-time"},"overtime":{"type":"integer","format":"int32"}}},"Worker":{"type":"object","properties":{"pid":{"type":"integer","format":"int32"},"name":{"type":"string"},"sex":{"type":"string"},"age":{"type":"integer","format":"int32"},"nationality":{"type":"string"},"idcard":{"type":"string"},"birthday":{"type":"string","format":"date-time"},"college":{"type":"string"},"address":{"type":"string"},"tele":{"type":"string"},"job":{"type":"string"},"jobtime":{"type":"string","format":"date-time"},"photo":{"type":"string"}}}}}}